+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enabled manually creating track or detection without imagery
 - Added default histogram bounds
 - Fixed bug that could occur when imagery is created with no frames
+- Added the ability to bulk label tracks
 
 ### Bug Fixes
 - Fixed bug where the indices of selected tracks were remembered such that after deleting selected tracks and loading